+++ go.mod
@@ -6,6 +6,6 @@ require (
 	github.com/jmoiron/sqlx v1.2.0
 	github.com/lib/pq v1.8.0
 	github.com/lissteron/simplerr v0.7.0
-	github.com/loghole/dbhook v0.2.2
+	github.com/loghole/dbhook v0.2.3
 	github.com/opentracing/opentracing-go v1.2.0
 )
